Experimental XXTEA cryptography for messages
Cryptographic operations appear in the right-click context menu. Just
highlight some text and right click it to encrypt or decrypt it. The
password is fed through a SHA-512 based iterative key derivation
function (not the best, but works with a decent amount of security.) Do
not use this for anything important because I cannot guarantee the
cryptographic properties of the KDF or XXTEA implementations.